requirements.txt
Django==4.2.6  
Pillow==10.4.0  
gunicorn==20.0.4  
django-crispy-forms==2.0.0  
crispy-bootstrap5==2023.10  
django-environ==0.8.1  
mysqlclient==2.1.0  
djangorestframework==3.14.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.0
python-decouple==3.8
billiard==4.2.1